Refactor simplification algorithm
Extract parts of algorithm to separate files.
