+++ cli/lib/simplification/simplify.js
@@ -1,206 +1,8 @@
 import { vec3 } from "gl-matrix";
-
-/**
- * symmetric 4x4 matrix
- */
-export class SymmetricMatrix {
-  constructor() {
-    // TODO: maybe use more specialized array?
-    this.data = new Array(10);
-    for (let i = 0; i < 10; i++) {
-      this.data[i] = 0;
-    }
-  }
-
-  static makePlane(a, b, c, d) {
-    const m = new SymmetricMatrix();
-    m.data[0] = a * a;
-    m.data[1] = a * b;
-    m.data[2] = a * c;
-    m.data[3] = a * d;
-    m.data[4] = b * b;
-    m.data[5] = b * c;
-    m.data[6] = b * d;
-    m.data[7] = c * c;
-    m.data[8] = c * d;
-    m.data[9] = d * d;
-    return m;
-  }
-
-  static add(a, b) {
-    const m = new SymmetricMatrix();
-    for (let i = 0; i < 10; i++) {
-      m.data[i] = a.data[i] + b.data[i];
-    }
-    return m;
-  }
-
-  det(m11, m12, m13, m21, m22, m23, m31, m32, m33) {
-    const m = this.data;
-    return (
-      m[m11] * m[m22] * m[m33] +
-      m[m13] * m[m21] * m[m32] +
-      m[m12] * m[m23] * m[m31] -
-      m[m13] * m[m22] * m[m31] -
-      m[m11] * m[m23] * m[m32] -
-      m[m12] * m[m21] * m[m33]
-    );
-  }
-}
-
-export class Triangle {
-  constructor(a, b, c) {
-    this.vertices = [a, b, c];
-    this.deleted = false;
-    this.isDirty = false;
-    // TODO: check if using normal from data would be more accurate
-    this.normal = null;
-    this.error = new Array(4);
-  }
-}
-
-export class Vertex {
-  constructor(x, y, z) {
-    this.position = vec3.fromValues(x, y, z);
-    this.q = new SymmetricMatrix();
-    this.isBorder = false;
-    this.tStart = 0;
-    this.tCount = 0;
-  }
-}
-
-export class Reference {
-  constructor(triangleIndex, vertexIndex) {
-    this.triangleIndex = triangleIndex;
-    // TODO: rename to triangleVertexIndex as it's not vertex index but index of vertex in triangle vertices array
-    this.vertexIndex = vertexIndex;
-  }
-}
-
-export const prepareData = (positionsArray, indicesArray) => {
-  const groupByThree = (arr) => {
-    return arr.reduce((agg, coordinate) => {
-      let last = agg[agg.length - 1];
-      if (!last || last.length === 3) {
-        last = [];
-        agg.push(last);
-      }
-      last.push(coordinate);
-      return agg;
-    }, []);
-  };
-
-  const positionsGrouped = groupByThree(positionsArray);
-
-  const vertices = positionsGrouped.map(([x, y, z], i) => {
-    return new Vertex(x, y, z);
-  });
-
-  const indicesGrouped = groupByThree(indicesArray);
-  const triangles = indicesGrouped.map(([a, b, c]) => {
-    return new Triangle(a, b, c);
-  });
-
-  return { triangles, vertices };
-};
-
-// TODO: maybe accept vec3?
-export const calculateVertexError = (q, x, y, z) => {
-  const m = q.data;
-  return (
-    m[0] * x * x +
-    2 * m[1] * x * y +
-    2 * m[2] * x * z +
-    2 * m[3] * x +
-    m[4] * y * y +
-    2 * m[5] * y * z +
-    2 * m[6] * y +
-    m[7] * z * z +
-    2 * m[8] * z +
-    m[9]
-  );
-};
-
-// error for one edge
-export const calculateError = (vertex1, vertex2) => {
-  // compute interpolated vertex
-  const qDelta = SymmetricMatrix.add(vertex1.q, vertex2.q);
-  const isBorder = vertex1.isBorder && vertex2.isBorder;
-  let error = 0;
-  let resultPoint = vec3.create();
-  // determinant is used to check if matrix is invertible
-  let det = qDelta.det(0, 1, 2, 1, 4, 5, 2, 5, 7);
-
-  if (det != 0 && !isBorder) {
-    // qDelta is invertible
-    // vx = A41/det(q_delta)
-    resultPoint[0] = (-1 / det) * qDelta.det(1, 2, 3, 4, 5, 6, 5, 7, 8);
-
-    // vy = A42/det(q_delta)
-    resultPoint[1] = (1 / det) * qDelta.det(0, 2, 3, 1, 5, 6, 2, 7, 8);
-
-    // vz = A43/det(q_delta)
-    resultPoint[2] = (-1 / det) * qDelta.det(0, 1, 3, 1, 4, 6, 2, 5, 8);
-
-    error = calculateVertexError(
-      qDelta,
-      resultPoint[0],
-      resultPoint[1],
-      resultPoint[2]
-    );
-  } else {
-    // det = 0 -> find optimal vertex along edge
-
-    const vertex3Position = vec3.create();
-    vec3.add(vertex3Position, vertex1.position, vertex2.position);
-    vec3.divide(vertex3Position, vertex3Position, vec3.fromValues(2, 2, 2));
-
-    const error1 = calculateVertexError(
-      qDelta,
-      vertex1.position[0],
-      vertex1.position[1],
-      vertex1.position[2]
-    );
-    const error2 = calculateVertexError(
-      qDelta,
-      vertex2.position[0],
-      vertex2.position[1],
-      vertex2.position[2]
-    );
-    const error3 = calculateVertexError(
-      qDelta,
-      vertex3Position[0],
-      vertex3Position[1],
-      vertex3Position[2]
-    );
-    error = Math.min(error1, error2, error3);
-    if (error1 === error) {
-      vec3.copy(resultPoint, vertex1.position);
-    } else if (error2 === error) {
-      vec3.copy(resultPoint, vertex2.position);
-    } else if (error3 === error) {
-      vec3.copy(resultPoint, vertex3Position);
-    }
-  }
-
-  return { error, point: resultPoint };
-};
-
-export const calculateTriangleError = (triangle, vertices) => {
-  triangle.vertices.forEach((vertexIndex, i) => {
-    const nextVertexIndex = triangle.vertices[(i + 1) % 3];
-    const { error } = calculateError(
-      vertices[vertexIndex],
-      vertices[nextVertexIndex]
-    );
-    triangle.error[i] = error;
-  });
-  triangle.error[3] = Math.min(
-    triangle.error[0],
-    triangle.error[1],
-    triangle.error[2]
-  );
-};
+import SymmetricMatrix from "./symmetric-matrix.js";
+import Reference from "./reference.js";
+import { calculateError, calculateTriangleError } from "./error-calculation";
+import { checkCollapseFlip } from "./check-collapse-flip";
 
 export const initializeData = (vertices, triangles) => {
   triangles.forEach((triangle) => {
@@ -305,57 +107,6 @@ export const compactTriangles = (triangles) => {
 };
 
 // TODO: add unit test
-// Check if a triangle flips when this edge is removed
-export const flipped = (
-  point,
-  index1,
-  vertex0,
-  deleted,
-  vertices,
-  triangles,
-  references
-) => {
-  for (let i = 0; i < vertex0.tCount; i++) {
-    const triangle = triangles[references[vertex0.tStart + i].triangleIndex];
-    if (triangle.deleted) {
-      continue;
-    }
-
-    const vertexIndex = references[vertex0.tStart + i].vertexIndex;
-    const id1 = triangle.vertices[(vertexIndex + 1) % 3];
-    const id2 = triangle.vertices[(vertexIndex + 2) % 3];
-
-    // delete check
-    if (id1 === index1 || id2 === index1) {
-      deleted[i] = true;
-      continue;
-    }
-
-    const d1 = vec3.create();
-    vec3.sub(d1, vertices[id1].position, point);
-    vec3.normalize(d1, d1);
-
-    const d2 = vec3.create();
-    vec3.sub(d2, vertices[id2].position, point);
-    vec3.normalize(d2, d2);
-
-    const dotProduct = vec3.dot(d1, d2);
-    if (Math.abs(dotProduct) > 0.999) {
-      return true;
-    }
-    const n = vec3.create();
-    vec3.cross(n, d1, d2);
-    vec3.normalize(n, n);
-    deleted[i] = false;
-
-    if (vec3.dot(n, triangle.normal) < 0.2) {
-      return true;
-    }
-  }
-  return false;
-};
-
-// TODO: add unit test
 // Update triangle connections and edge error after an edge is collapsed
 export const updateTriangles = (
   i0,
@@ -452,7 +203,7 @@ export const simplify = (vertices, triangles, targetTriangles = 200) => {
 
           // don't remove if flipped
           if (
-            flipped(
+            checkCollapseFlip(
               point,
               vertexIndex1,
               vertex0,
@@ -465,7 +216,7 @@ export const simplify = (vertices, triangles, targetTriangles = 200) => {
             return;
           }
           if (
-            flipped(
+            checkCollapseFlip(
               point,
               vertexIndex0,
               vertex1,
