+++ cli/lib/simplification/simplify.test.js
@@ -1,14 +1,9 @@
 import { vec3 } from "gl-matrix";
-import {
-  Vertex,
-  SymmetricMatrix,
-  prepareData,
-  initializeData,
-  calculateError,
-  calculateVertexError,
-  getReferenceList,
-  compactTriangles,
-} from "./simplify";
+import { initializeData, getReferenceList, compactTriangles } from "./simplify";
+import Vertex from "./vertex.js";
+import SymmetricMatrix from "./symmetric-matrix.js";
+import { prepareData } from "./prepare-data.js";
+import { calculateError, calculateVertexError } from "./error-calculation";
 
 describe("simplify", () => {
   const positions = [0, 0, 0, 1, 0, 0, 0, 1, 0];
